refactor: group report generation files into lib/reporting/
