+++ src/components/vulnerability-scatterplot.tsx
@@ -25,7 +25,7 @@ import {
   transformScansToChartData,
   generateChartKey,
   type ScatterplotDataPoint,
-} from "@/lib/scatterplot-data"
+} from "@/lib/reporting/scatterplot-data"
 
 export function VulnerabilityScatterplot() {
   const router = useRouter()
