+++ src/lib/reporting/scatterplot-data.ts
